fix render and add gropping
