+++ .github/actions/tests/action.yml
@@ -31,20 +31,32 @@ runs:
         echo "always-auth=true" >> .npmrc
         echo "::endgroup::"
 
+        echo "::group::Installing and updating dependencies"
         cd "packages/${{inputs.package}}"
         yarn install
         npm run deps --if-present -- --skip-commit
+        if [ ! -z "${{inputs.install}}" ]; then yarn add --dev ${{inputs.install}}; fi
+        echo "::endgroup::"
 
         if [ ! -z "${{inputs.links}}" ]
         then
-          npx link --if-present -- --include ${{inputs.links}} --build --install --depth ${{inputs['linking-depth']}}
+          echo "::group::Linking dependencies"
+          npx link --include ${{inputs.links}} --build --install --depth ${{inputs['linking-depth']}}
+          echo "::endgroup::"
         fi
 
-        if [ ! -z "${{inputs.install}}" ]; then yarn add --dev ${{inputs.install}}; fi
-
+        echo "::group:Building package"
         npm run build --if-present
+        echo "::endgroup::"
+
+        echo "::group::Setting up test environment"
         npm run setup --if-present
+        echo "::endgroup::"
 
+        echo "::group::Running tests"
         if [ ! -z "${{inputs.type}}" ]; then npm run test:${{inputs.type}}; else npm run test; fi
+        echo "::endgroup::"
 
+        echo "::group::Sending reports"
         if [ "${{inputs.sandbox}}" = "sandbox" ]; then npm run report --if-present -- --sandbox; else npm run report --if-present; fi
+        echo "::endgroup::"
