+++ packages/eyes-playwright/package.json
@@ -69,9 +69,7 @@
     "@applitools/api-extractor": "1.1.1",
     "@applitools/scripts": "file:../scripts",
     "@applitools/sdk-coverage-tests": "2.3.11",
-    "@applitools/sdk-fake-eyes-server": "2.1.1",
     "@applitools/sdk-release-kit": "0.13.0",
-    "@applitools/sdk-shared": "0.8.10",
     "@applitools/test-utils": "file:../test-utils",
     "@typescript-eslint/eslint-plugin": "^4.15.1",
     "@typescript-eslint/parser": "^4.15.1",
