+++ packages/test-utils/index.js
@@ -1,0 +1,4 @@
+exports.getTestDom = require('./src/get-test-dom')
+exports.getTestInfo = require('./src/get-test-info')
+exports.parseEnv = require('./src/parse-env')
+exports.setupEyes = require('./src/setup-eyes')
